backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "boto3>=1.42.11",
    "fastapi>=0.124.4",
    "google-generativeai>=0.8.6",
    "mangum>=0.19.0",
    "openai>=2.13.0",
    "pypdf>=6.4.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "python-pptx>=1.0.2",
    "uvicorn>=0.38.0",
]
